host-app/requirements.txt
# Tact Host Application Dependencies
# Install with: pip install -r requirements.txt

pyserial>=3.5